New method Alias::owns ().
The method is very similar to Alias::exists () but unlike the latter
one it takes user_id as a second parameter. This allows us to check
not only that the alias exists but is also a property of a specific user.

Method Alias::delete () requires a second parameter user_id. In theory
the parameter could be omitted as a call to Alias::delete () is always
preceeded by a call to Alias::owns (). In reality shit happens so consider
it a safety precaution.