Do not check existence of keyrings.
gpg does it for us and if possible, creates them before issuing a desired
command.

Also, the scripts that call gpg command, that reads data from stdin,
no longer use 'cat' to pipe a content into it. gpg is smart enough
to do that on its own (possible with a little help of --batch option).